fix: improve firefox init script
Use no remote cause apparently not having that causes some issues?
+ early return check & wait for proc to die

+++ firefox/init_firefox.sh
@@ -1,15 +1,21 @@
 #!/usr/bin/env bash
 
-firefox --headless about:blank &>/dev/null &
+profiles_ini="${HOME}/.mozilla/firefox/profiles.ini"
+
+# Profiles already exist
+[[ -f "$profiles_ini" ]] && exit 0
+
+# Launch firefox to generate profiles
+MOZ_NO_REMOTE=1 firefox --headless --no-remote about:blank &>/dev/null &
 ff_pid=$!
 
 # Wait for the profiles.ini but stop if Firefox dies or time out
 tries=0
-while [[ ! -f "${HOME}/.mozilla/firefox/profiles.ini" ]]; do
-    kill -0 "$ff_pid" 2>/dev/null || exit 0
-    (( tries++ >= 30 )) && break
+while [[ ! -f "$profiles_ini" ]]; do
+    kill -0 "$ff_pid" 2>/dev/null || break
+    (( tries++ >= 60 )) && break
     sleep .5
 done
-sleep .5
 
 kill "$ff_pid" 2>/dev/null
+wait "$ff_pid" 2>/dev/null
